Fix notification logic for signals in BlockListImpl
The logic for calling onItemAdded and onItemRemoved signals when
setting a new list of block items using BlockListImpl::setItems
used to be broken.

This commit fixes and documents the correct signal notification
behavior

Test-Information:

Added a unit test which verifies the notification behavior in case
of added block list items, removed block list items and a complete
change of the block list.

Change-Id: I3061545e25ddfc2d9d1a3c987045a58e5c9230ac