Require a node-part for JIDs entered in 'Add contact…'-dialog
While it is possible to add domain only JIDs to a user's roster
this is rarely the intention of a user. This patch adds the
requirement of a node-part for the address entered in the dialog
and presents a warning if the node-part is missing.

Test-Information:

Verified that a warning triangle is shown if there is no node-part
in the user entered address.

Change-Id: I38dd0b34ca7cb19b38bb7e1c457f7a8a9ef84028