Ignore DND drops of JIDs for contacts already in the conversation
Swift allows dropping contacts from group chats or the roster
on a chat window to invite them or start an impromptu chat.
With this commit Swift will ignore drops which only contain
JIDs that are already part of the conversation.

Test-Information:

Tested the described behavior with roster and room contact
drags in anonymous and non-anonymous rooms.

Change-Id: I2f06082ea4bc1140210f9f1a165bdf276a130273