Fix hang during file-transfer preparation with missing proxy
The code used to call the onProxiesFound signal only if a
proxy was found. In case of missing S5B proxy at the users
server, the file-transfer preparation would hang.

Now the code wants until the discovery phase is done and then
calls the onProxiesFound signal with the list of discovered
proxies. In case of missing S5B proxy server the signal is
called with an empty list so the file-transfer flow can
continue.

Test-Information:

Ran FileTransferTest integration test and manually tested a
file-transfer on a server without S5B proxy.

Change-Id: I31d3cc08fe6453b5cdfe6be286f884a920470d28