Only bind full JID to ChatController on non-empty or typing messages
With this commit a ChatController is only bound to a full JID
of a contact, if the incoming message of the contact has a
non-empty body or is a 'typing' chat state notification.

This avoids random binding to an arbitrary resource of a contact,
that has multiple online resources and responds with delivery
receipts.

Test-Information:

Tested with a conversation of a one resource account (A) to a
two resource account (B). Sending two messages to B, the first
used to go to both resources, then A would bind to the full
JID and the second message would only go to one resource.

With this fix all messages go to both resources, i.e. are
send to the bare JID, until one resource of B replies. This
binds the controller at A to the full JID of the reply of B.

Change-Id: I8d9321a4226ab798e1196351ad087990d5dff8c3