Add additional unit tests for Chat <-> JID binding
The recent commit 582ca91 changed the behavior the binding
between chat sessions and the full JID of a chat contact.

This commit adds additional unit tests verifying the behavior
of implemented in 582ca91. This means a chat session binds to
the full JID of an incoming message only if the message has
a non-empty body text or is a 'typing' chat state notification.

It also tests the rebind to another resource if a 'typing'-CSN
or a body message is received from that resource.

Test-Information:

All unit tests pass on OS X 10.10.5.

Change-Id: Id3f25d8a3ff78c407ed4b2a97bd421dc4aa58688