Add support for client certificate authentication in BOSH
This allows to authenticate using SASL EXTERNAL over BOSH
using a client TLS certificate for the HTTPS connection of
the BOSH channel.

The implementation also enforces the HTTPS server certificate
of subsequent BOSH connections not to change.

This commit also removes TLSConnection and TLSConnectionFactory
as no code is using them.

Test-Information:

Tested against M-Link 16.3v6-0 on Debian 7.9 and Swift on
OS X 10.10.5. Verified working client certificate authentication.
Verified Swift not falling back to password-based authentication,
in case EXTERNAL is not allowed by the server over BOSH or
the client certificate is invalid.

Change-Id: Ia96bcac27cac9fc9261ed847c82c6328307bfbd1