Change bare JID presence lookup code to ignore priorities
Before presence handling code was handled by both, the
ContactRosterItem in Swift and the PresenceOracle in Swiften.
The ContactRosterItem also considered the presence priority
for deciding what presence to show for a bare JID.

With this code all full or bare JID presence requests are
finally handled by the PresenceOracle. For bare JIDs it is
looked up to a presence of one of the available resources of
that JID regardless of the priorities.

Test-Information:

Adjusted tests according to above description and documentation
in PresenceOracle.

Change-Id: I972a4574f476cdf4d4b5593a035eb1c25ef2f8ba