Making virtual calls in constructor of ChatControllers explicit
Virtual calls in ctors/dtors are tricky as it is not clear
what method is going to be called. By changing the code to
call the explicit method it is obvious.

This issue was raised by Clang Analyzer.

This commit also adds missing SWIFTEN_OVERRIDE annotations.

Test-Information:

Still compiles, tests pass and Clang Analyzer warning so gone.

Change-Id: I8096994277ad32933f992b9a11079c761d3ec619