Fix sluift API for clearing certificate
Before this patch it was calling the constructor and creating certificate
object with empty string. This was causing tests to fail when client was
set with null certificate.
This object fixes it by setting the certificate object to Null if the
certificate file is empty.

Test-information:
Test cases using sluift with this API that set null certificate now pass

Change-Id: Id75e22b58afebb2db463a7a23a4ca0686ef6eb12