Workaround QTabBar early eliding bug on OS X
The issue is visible when having two chat windows with the
same label. One of them is elided even though enough horizontal
space is available.

Test-Information:

Tested multiple chats including some with the same tab
label.
Tested many open chat windows overflowing the horizontal space
of the tab bar to test scroll behavior on overflow.

Verified that the code stops the early eliding bug on
OS X 10.10.5 with Qt 5.4.2.
Verified that workaround is not needed on Debian 8.2 with
Qt 5.4.2.

Change-Id: I2dc0d417fb6f402dda2f7575a83ca3faf4eb63cf