Disconnect signals from in-flight IBB request
Canceling an IBB transfer in-flight caused an use-after-free
reported by ASAN. With this fix we keep a reference of the
current request around to be able to disconnect from its
signals on cancel.

Test-Information:

Transferred a file with Swift using the IBB method and canceled
the transfer. The previously error reported by ASAN is gone.

Change-Id: I240d3dbb59cddb6b91d49f268595a89ac8805f72