Fix IPv6 DNS resolution issues on Windows
Use values instead of define names for _WIN32_WINNT and
NTDDI_VERSION defined needed for Windows.h configuration.
Using the names boost fails to correctly detect getaddrinfo()
support on Windows.

Only run IPv6 related test cases in DomainNameResolverTest on
Windows, if test_ipv6=1 is passed to the scons arguments.

This is because on Windows getaddrinfo() will not return
IPv6 related results when called with the AF_UNSPEC hint,
unless the Windows host has global IPv6 connectivity.

Changed the BoostConnectionTest to time out and not endlessly
wait on a response from the remote host.

Test-Information:

Ran the following test configurations:

* (SUCCESS) On Windows 8 with HE.net IPv6 tunnel to provide
  full IPv6 connectiviy:
  scons.bat test=system test_ipv6=1 Swiften/QA/NetworkTest

* (SUCCESS) On Windows 8 with HE.net IPv6 tunnel to provide
  full IPv6 connectiviy:
  scons.bat test=system Swiften/QA/NetworkTest

* (EXPECTED FAIL) On Windows 8 with no IPv6 connectiviy:
  scons.bat test=system test_ipv6=1 Swiften/QA/NetworkTest

* (SUCCESS) On Windows 8 with no IPv6 connectiviy:
  scons.bat test=system Swiften/QA/NetworkTest

Change-Id: I5adcd28e09e22acf61f7cca40b614e71df75dd70