Remove warning about undefined QTDIR environment variable
It used to be a common way to use the QTDIR environment
variable to tell SCons the location of the Qt installation
on the system.
Swift however uses the 'qt' SCons config variable for the
location and the 'qt5' variable for the version of Qt to use.
On Unix/Linux systems with pkg-config, SCons will use it to
detect the Qt installation and its flags.

The QTDIR environment variable is rarely used and tested so
the warning about missing environment variable is removed.

Test-Information:

Tested on Debian 8.

Change-Id: I2cd793f12b8607743db2848765011cc719b106aa