Stop timer during clean up of WhitespacePingLayer
AddressSanitizer reported a heap-buffer-overflow in
WhitespacePingLayer::handleTimerTick() which happened during
multiple restarts of a XMPP server. Under the assumption
that it was caused by not correctly stopping the timer, the
timer is now explicitly stopped if still active at destruction.

Test-Information:

Unable to reproduce the ASAN report. Unit tests still pass.

Change-Id: Ia0b7c3b613688750c4ce0ad40d759a0db4a52791