Ensure that BoostTimers can't fire after stop()
Code review makes me think that if the timer fires at the same time as
stop() is called, it would be possible for the clearing of events from
the loop to happen before the event is put into the loop, leading to
the sort of crashes we've seen with timers firing and access exceptions.
I can't reproduce those to know if this fixes them, and I'm not even sure
I'm not being dense thinking this patch fixes a real issue.

Test-Information:
Unit tests pass

Change-Id: I76337d5556a9c3902d5c2f4da754ae657810d436