Treat cert verify errors as non-fatal in OS X TLS backend
Our TLS backends need to tread TLS verification errors, e.g.
outdated certificate, untrusted CA, non-matching host, etc.,
as non-fatal, so the application can apply custom key
pinning verification or similar.
This patch changes the OS X SecureTransport backend to behave
accordingly and adjusts the CertificateErrorTest to mirror
this behavior.

This commit also fixes a double-free in
SecureTransportCertificate.

Test-Information:

Connected to a host with an untrusted CA and non-matching
domain in the certificate and was prompted with the Swift
certificate trust dialog on OS X 10.11.3.

Swiften/QA/TLSTest run successfully on OS X 10.11.3.

Change-Id: I4c8ce2178540d79a5f328e2e0558d4deb4295134