Fix file-transfer UI replace after presence change glitch
Forgot to reset the lastWasPresence_ flag when file-transfer
UI is added to a chat view. The flag is used to amend the
existing presence change log in a chat view in case of
consecutive presence changes.

Test-Information:

Transfer a file to a Swift user and go offline afterwards.
Without this patch the file-transfer UI is replaced by
a presence change message. With this patch a new presence
change message is added instead.

Change-Id: I54a5cfd398bd4ba79eb15a1b14b08db3377233a4