Fix using external OpenSSL on Mac OS X
This commit changes our TLS backend configuration behavior.
It introduces the tls_backend Scons argument, which defaults
to the native backend, but can also explicitly set to
'openssl' to use a project external OpenSSL installation or
'openssl_bundled' to build and configure with the OpenSSL
source in the 3rdParty directory.

Test-Information:

Tested on OS X 10.11.3.

./scons ends up using Secure Transport.

./scons tls_backend=openssl \
	openssl_include=/usr/local/opt/openssl/include \
	openssl_libdir=/usr/local/opt/openssl/lib \
	openssl=/usr/local/opt/openssl successfully builds with
	openssl from Homebrew.

./scons tls_backend=openssl_bundled successfully builds with
	OpenSSL in 3rdParty.

Change-Id: I4fb0ef9d197609afe793554f86e54fe67fc1cab5