Disable Hunspell support by default as it is not finished yet
Our Hunspell support has not been finished and thoroughly
tested yet. This commit disabled building with Hunspell
by default, even if the dependencies for it are installed
on the system.

Remove libhunspell-dev from the build depends in Debian
packaging.

Test-Information:

Tested on Debian 8.
Tested with and without the hunspell_enable=True flag.
With the spellchecking options are shown and however spell
checking is not stable.
Without it the spellchecking options are not shown and
spellchecking is disabled in general, as expected.

Change-Id: I100582852cbe128149fdb2e2a28d4ae407398be6