Improve profile editing UX when vCard is not supported
Previously, when editing your own vCard and the server
responded with an error, the user would indefinitely see a
spinner.
With this commit, the user will see an error message instead.

Test-Information:

Tested on OS X 10.11.3, with a XMPP server with enabled and
disabled vCards. The enabled case still works as expected and
in the disabled case the error message is shown, instead of
the indefinite spinner.

Change-Id: Ic9167ee633a2f9a9fa3b520f6067dc2c94857c07