Fix crashes in spellchecking code in case of broken backend
If Swift uses Hunspell as spellchecking backend its
configuration can fail, because a dictionary is missing
or a wrong path was specified.
Instead of crashing, spellchecking will be disabled and a
warning is printed to the terminal.

Test-Information:

Tested on Debian 8 with Qt 5.3.2.
Tested enabling/disabling of spellchecking in the UI,
selecting different dictionaries and broken dictionaries.

Change-Id: Ib6d73ed3d7a6a3701410b0f6dc983d41e807df82