Refactored keyword highlighting
This commit changes the ChatWindow/ChatView APIs to not pass
highlights actions as additional parameters but instead they
are now part of the ChatWindow::ChatMessage and its parts.

This allows the controllers to do highlighting in one single
place and play sound actions on the highlighted message in
a single place.

On a highlighted message only unique sounds are played and
they are played in sequence of the rules that matched
the message.

Test-Information:

Adjusted the existing unit tests accordingly. Added unit
tests that check reduplication of highlight action sounds
and that the sound actions are emitted correctly.

Manually verified that highlight sound actions with and
without duplicated sounds are audible on OS X 10.11.3.

Change-Id: I68c88e0d285d79d87b2997ed29d92b140480b394