Fix warning coloring of chat input field
The previous stylesheet was too generic, as it applied on
all widgets and sub widgets, including a potential context
menu. This fix will only color alerts and the text input
in the warning color. Potential context menus are colored
according to the application style.

Test-Information:

Tested alert colors and text input color on OS X 10.11.3
with Qt 5.4.2.

Change-Id: I0d1f704f4eacddf41befb7dfbc8bbf7ef28f88ce