Replace chat view theme with one based on the new design
This also removes the old chat theme resources and some code
in QtWebKitChatView that was required for compatibility to
Adium-style themes.

The new code uses a CSS style in the header to change the
font size and does not iterate the whole DOM tree itself
anymore.

Added new resources for failed and successful asks.

Test-Information:

Tested MUCs, PMs, message correction and file-transfers with
the new chat theme.

Change-Id: If922a972c658189444e60a7b00e5e5e96661620d