Break long words (e.g. URLs) in chat view to prevent overflows
This prevents a horizontal scroll bar as long as there is
space for the message block. If it is resized to a size that
only allows horizontal space for the dark side block, then
there will still be a horizontal scroll bar.

Test-Information:

Tested in various chat rooms by resizing the splitter between
chat view and occupant list.

Change-Id: Ibdf6a9554adcbd7580cb9b40eaffff6279a184b0