Fix delayed login/roster resize on window resize in netbook mode
If you resize the window in netbook mode, the splitter inside
only resizes to the new geometry if it is clicked or moved.
This is due to overriding resizeEvent hook of QSplitter in
QtSingleWindow but not notifying QSplitter's resizeEvent hook
in the new method.

Test-Information:

Tested on OS X 10.11.4 with Qt 5.4.2.

Opened Swift in netbook mode and resize the window. The login
screen is resized as the window is resized. The login screen
used to only resize if you moved the splitter.

Change-Id: I7715e1cdcb4c0e672e233a9f76fad3ef5fe70d25