Use FeatureOracle to detect file-transfer support in roster
The FeatureOracle provides tri-state feature lookup
functionality for bare JIDs. It returns Yes if a feature is
supported by all resources of the bare JID, Maybe if some
support it, and No if none of the resources support it.
If passed a full JID, it returns the specific features supported
by that end-point.

Sending a file to a bare JID, will send a file to the resource
of the bare JID with the highest availability by presence, show
status and priority and which supports the features required
for a Jingle file-transfer.

Test-Information:

Added unit test verifying new behavior.

All tests pass on OS X 10.11.6. Added new unit tests for
FeatureOracle.

Manually verified that the roster and chat window both use
the same mechanism to detect support for file-transfers.

Manually verified that file-transfers via the contact list
goes to already bound full JIDs if there is an existing
ChatController.

Change-Id: I0175ac42ecb73f1d54f9c96ffbba773eb5e24296