Allow to invite multiple contacts to a chat room at once
There was a bug that prevented to drag multiple contacts
in the invite dialog list. The restriction for impromptu
conversation rooms is now only applied if no local MUC service
is found and the dialog is a 'Chat to Users' dialog.

Test-Information:

Tested that you can now drag multiple contacts into the
invite dialog.

Tested on OS X 10.11.4 with Qt 5.4.2.

Change-Id: I50b0bb191570f32cfa054e415c993b1fb57073ad