Add minidump analysis python script for Windows crash dumps
This script takes a Swift version and minidump file path as
arguments and will checkout the corresponding source, fetch
and extract the corresponding installer, fetch and extract
the corresponding debug symbols to finally run the Windows SDK
command line debugger (cdb) to write a human readable stack
trace and further info to a txt file.

Test-Information:

Tested it successfully on Windows 10 with various minidumps.

Change-Id: I5915255dfb23e6b45c1273a3c1c42cc83125df57