Changing default avatar to a blue circled swift
The old default avatar does not work well with the new
chat view design which uses circled avatars.

This also uses the vector avatar internally in Swift and
has Qt and the chat view engine scale it to the required
size.

Test-Information:

Tested on OS X 10.11.5 with non-compact chat views and MUC
conversations.

Change-Id: I82580ecb91fc81f4fe9535aa9d3b214b720a40f3