Add presence to manually added contacts in 'Start Chat…' dialog
This fixes a bug where manually added contacts, i.e. the user
entered the address and hit the 'Add to list' button, were
missing the correct presence in the contacts list in the
'Start Chat…" dialog.

Test-Information:

Tested on OS X 10.11.5 with Qt 5.4.2.
Tested adding a roster member via the drop down suggestion and
via manually entering the JID and clicking the 'Add to list'
button. In both cases the correct presence was shown.

Added a test verifying that empty JIDS are invalid.

Change-Id: Idb18e4ad2b5c0afb48461796f64e8f11f02a24fd