Fix Swift UI sometimes getting stuck during login
In some cases during the login the helper QEventLoop in
QtWebKitChatView takes over the event handling of the
application and never retirms from it. Qt events are still
handled in this case but any other new events are only queued
up. The user would still have a responsible UI, but all network
traffic or other non-Qt events would not be handled correctly.

Test-Information:

Without the patch Swift would get stuck during login in 5 of
10 runs. With this patch it did not get stuck anymore in 10
runs.

Tested on OS X 10.11.5 with Qt 5.5.1.

Change-Id: Ie17a8e01cc9c52cc784e223974ec21e604dcccab