Draw usernames in contact list in dark gray
It now uses the same dark gray tone as the chat view uses
for its text.

Moved some constants to named variables for readability.

Test-Information:

Compact and non-compact contact lists, in the roster and MUC
roster and the chats view all show the first line drawn in
a dark gray instead of the hard black.

Change-Id: Ief36fd4f27d30ace61f04d50e846db571ab84a4e