Add support for signing the resulting MSI installer on Windows
Test-Information:

Created a custom CA and added it as trusted to the system.
Created a signing certificate and passed it to scons as
described. The resulting installer does not show a red UAC
dialog anymore during installation. Instead the publisher from
the certificate is shown in a blue UAC dialog.

Change-Id: Ie4043520f6d45ec2e7aad712441a928cb423b0de