Optionally create GPG signature for generated tarball
If the gpg tool is found in the PATH and a key exists for
Swift's signing identity, package.sh will generated a GPG
signature (.asc) file alongside the tarball.

Test-Information:

Created a new GPG key with Swift's signing identity,
ran package.sh script and correctly verified the generated
signature with GPG.

Tested both with and without --nosign flag.

Tested on Debian 8.5 with GPG 1.4.18.

Change-Id: I7e6c8b6a7c900b68e53ece97a59eba33a2de565d