Support dropping tabs on tab widgets
Previously users could only drag and drop tabs on other tab
bars. With this change the user can simply drop the tab on
another tab widgets. This works on all QtTabWidgets that have
a QtDNDTabBar set, regardless of their number of child widgets.

This works by forwarding dragEnverEvent() and dropEvent() calls
on the QtTabWidget to the corresponding QtDNDTabBar.

Tabs dropped on the tab widget instead of the tab bar are
added at the end of the tab bar.

Test-Information:

Tested with Qt 5.5.1 on OS X 10.11.6 and Windows 7.

Change-Id: Ie73e02de24472eab2d20a89c937fb6630b1ef7b8