Improve chat view font handling
This changes the minimal font scaling to 0.8em, so that the
font size in the chat view can be smaller than the font size
in the rest of the application.

At 1.0em the font in the chat view has the same size as the
rest of the application.

Also changed fixed pixel sizes in main.css to relative
percentage sizes.

Test-Information:

Verified the font size in the chat view is the same as the
font size for the contact lists at 1.0em on Windows and OS X
with Qt 5.5.1.

Change-Id: Ib625f11cea0aacd31edee27c2dc89df5e2cbd1c4