Improve readability of message receipt icons
This changes the icons for delivery success and delivery
failure to vector graphics in SVG format.

Added a slight white circle around the graphics for better
readability on backgrounds with low contrast.

Test-Information:

Tested with Qt 5.4.2 and Qt 5.5.1 on OS X 10.11.6 on retina
and non-retina systems.
Noticeable better rendering due to SVG over PNG. Noticeable
but not attracting attention white circle.

Change-Id: I3c9be99854a300a4e40161a74964514aa47d292a