Add simple versioning to QtScaledAvatarCache
The cache for each version is stored in a dedicated directory
that exists next to the raw avatar images.

This allows running different Swift versions which use
different versions of the QtScaledAvatarCache.

Test-Information:

Tested on OS X 10.11.6 with Qt 5.5.1 and Windows 8 with Qt
5.5.1.

Verified that the avatar paths changed to the new versioned
cache path in the web views.

Change-Id: I12cb46c4d6750ae7ad82c719eef290f721d1d131