Improve visibility of currently focused chat input in trellis mode
The default focus indicator on the text input fields in chat
windows becomes harder to notice in trellis mode as there are
many similar looking chat windows visible at the same time.

This change increases the visibility of the focus indication
border on Windows and will enable the
standard blue focus halo on OS X for focused chat inputs
in trellis mode.

Test-Information:

Tested with Qt 5.5.1 on OS X 10.11.6 and Qt 5.5.1 on Windows 8.

Change-Id: If9fe9edea6fef292bb99eabbb125c7a9ec20dcc2