Improve roster item readability for selected items
Due to the style independent fixed text color for the status
message of non-compact roster item, it was barely readable on
Windows 7. This commit has the color for the status message
in non-compact roster items depend on the used text color
for the name. It uses a brightness adjusted highlight text
color for the second line.

Test-Information:

Tested on OS X 10.11.6 with Qt 5.5.1 and Windows 8 with Qt
5.5.1.

Change-Id: I41beb0f3eaede3f9413a39662c213a4e904bdd69