Verify message IDs during last message correction
Previously we did not check the ID in the replace tag against
the ID of the last message from that JID because some MUC
components change the message ID.

In case the ID of the last message and the ID in the replace
tag do not match, the message is simply treated as a normal
message.

Test-Information:

Added unit tests to verify the new behavior and adjusted
existing test cases for new behavior.

Added test cases to ChatsManagerTest.cpp that test verification
of replacement IDs for 1-to-1 chats and test non-verification
of replacement IDs for MUC.

All tests pass on OS X 10.11.6.

Change-Id: I85b1d2138b056b445a663f3ee3ab89a56cef4a2a