Provide explanation for empty trellis cells to user
This shows a short descriptive text in empty trellis cells,
that describes the current state to the user and suggests
actions on how to get rid of the empty trellis cell, e.g.
by changing the layout or moving chats to the empty cells.

Test-Information:

Tested with Qt 5.5.1 on macOS 10.12.

Change-Id: Ibc7d3a327cf32b8a21304b37e79534773b5d3761