Fix building Swift on 32-bit Linux distributions
Some distributions ship Qt configured so it requires your
application to be build with the -fPIC flag. This adds a
SCons check for this so the flag is automatically added if
required.

Test-Information:

Successfully build Swift and ran unit tests on Ubuntu 16.04
(32 bit and 64 bit) and Debian 8.6 (32 bit and 64 bit).

Change-Id: I8016bcb43fa36a3becf946ec179d74cdceb1d147