Scale QWebView DPI resolution to match desktop DPI resolution
Test-Information:

Tested that at 1em scaling, the default, the main text in the
chat style has the same height as the nicknames in the roster.

Tested with Qt 5.4.2 on macOS 10.12 on a retina system,
with Qt 5.4.2 on macOS 10.12 on a non-retina system, and
Windows 8 with Qt 5.7.0 with QtWebKit Technology Preview 4.

Change-Id: I10701c411d9f787bf497eb6aab208a0f3fda621c