Make avatar size relative to font size
A recent change to web view scaling to have equal font sizes
in and outside the web view cause the pixel sized avatars
to be half the usual size. With this fix the avatar is as
high as The the nickname plus the first line of the first
message.

Test-Information:

Tested on macOS 10.12 with Qt 5.7.0 and QtWebKit TP4.

Change-Id: I13e3c83fdb00e0dc9079c3ad5d6a380f6db5a499