Reduce width of sidebar in chat theme; move time inline
The moves the timestamps from the sidebar into the message
block to the right, after the nickname, floating right-aligned
at the top of a message.

This reduces the width of the sidebar.

Test-Information:

Tested with Qt 5.7.0 and QtWebKit TP4 on macOS 10.12.
Tested MUC and 1-to-1 chat views. Tested last message
correction UI behaviour.

Change-Id: Ibd899e9017ac5654b3eb3fe909c9be1524d4b50b