Sluift: Added Carbons element convertors
Sluift was showing payloads coming from a carbon copied message as
a series of dom payloads, which was not really usable.

With these ElementConvertors, the payload is translated into a
much more sensible table

Also a minor update to both presence/message events, which now
will report the 'to' as well. Seemed handy to have.

Test-information:

Carbons sent/received messages are formated into a meaningful table
now.

Change-Id: I5b8943636e09e5377bde76d16970c01df29164d6