Improve string to HostAddress conversion API
Previously HostAddress had a constructor which allowed
initialisation via a std::string. This initialisation can
fail and this is heavily used for checking whether a string
is a valid IP address.

This constructor is removed in this commit and replaced by
a static method HostAddress::fromString, taking a string and
returning an optional HostAddress. This clearly communicates
that the conversion can fail.

Test-Information:

./scons test=all passes on macOS 10.12.1.

Change-Id: Idaafee6f84010ce541c55f267ac77ad6ac8f02b4