Show MUC PM item-not-found error in MUC PM window
Previously, if one would send a MUC PM to a user that has
left a room the error response from the MUC would be shown
in the MUC room.
Now the error will show in the MUC PM window, if the MUC PM
full JID has a ChatController.

Test-Information:

Added unit test verifying new behaviour. Manually verified that
the error is shown in the MUC PM window instead of the MUC room
window.

Change-Id: I1b259d5eee9e22217bbe7e5c09294d2166a77895